packages/gg_api_core/pyproject.toml
[project]
name = "gg-api-core"
version = "0.1.0"
description = "Core functionality for GitGuardian MCP servers"
authors = [
    { name = "GitGuardian", email = "support@gitguardian.com"},
]
maintainers = [
    {name = "GitGuardian", email = "support@gitguardian.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "License :: OSI Approved :: MIT License",
    "Environment :: Console",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Security",
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "fastmcp>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.0",
]
license = {text = "MIT"}

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
